+++ lib/ReactViews/Story/StoryBuilder.tsx
@@ -258,6 +258,21 @@ class StoryBuilder extends Component<
     this.props.viewState.terria.stories = sortedArray;
   }
 
+  @action
+  updateShareDataStringSize() {
+    const shareData = getShareData(this.props.viewState.terria);
+    const shareDataString = JSON.stringify(shareData);
+    this.props.viewState.setShareDataStringSize(shareDataString.length);
+  }
+
+  componentDidMount(): void {
+    this.updateShareDataStringSize();
+  }
+
+  componentDidUpdate(): void {
+    this.updateShareDataStringSize();
+  }
+
   componentWillUnmount() {
     this.clearRecaptureSuccessTimeout?.();
   }
@@ -476,6 +491,15 @@ class StoryBuilder extends Component<
   render() {
     const { t } = this.props;
     const hasStories = this.props.viewState.terria.stories.length > 0;
+    const shareDataSize = this.props.viewState.shareDataStringSize;
+    const shareMaxRequestSize =
+      this.props.viewState.terria.shareDataService?.shareMaxRequestSize;
+    const shareMaxRequestSizeBytes =
+      this.props.viewState.terria.shareDataService?.shareMaxRequestSizeBytes;
+    // Disable the warning if map owners use custom server that does not return shareMaxRequestSize:
+    const shareDataTooLong = shareMaxRequestSizeBytes
+      ? shareDataSize > shareMaxRequestSizeBytes
+      : false;
     return (
       <Panel
         ref={(component: HTMLElement) => (this.refToMeasure = component)}
@@ -507,13 +531,23 @@ class StoryBuilder extends Component<
           </Text>
           <Spacing bottom={2} />
           <Text medium color={this.props.theme.textLightDimmed} highlightLinks>
-            {t("story.panelBody")}
+            {t("story.panelBody") +
+              (shareMaxRequestSize
+                ? " " + t("story.panelBodyCapped", { shareMaxRequestSize })
+                : "")}
           </Text>
           <Spacing bottom={3} />
           {!hasStories && this.renderIntro()}
           {hasStories && this.renderPlayShare()}
         </Box>
         <Spacing bottom={2} />
+        {shareDataTooLong && (
+          <Box paddedHorizontally={2}>
+            <Text small color={this.props.theme.textWarning} highlightLinks>
+              {t("story.storiesTooLong")}
+            </Text>
+          </Box>
+        )}
         {hasStories && this.renderStories()}
         {this.state.editingMode && (
           <StoryEditor